% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.arlingtoncounty.parks-package.R
\docType{package}
\name{dc.arlingtoncounty.parks-package}
\alias{dc.arlingtoncounty.parks}
\alias{dc.arlingtoncounty.parks-package}
\title{dc.arlingtoncounty.parks: Social Data Commons: Environment, Land Use: Arlington County Parks Access}
\description{
Allows user to easily get Arlington County Parks Access data from the SDAD Data Commons. Data can be provided directly or in a file.
}
\author{
\strong{Maintainer}: Eric Oh \email{eo9tv@virginia.edu}

Authors:
\itemize{
  \item Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia
}

}
\keyword{internal}
